+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-
+### Changed
+- Use LogService from ZServices package instead of ZUtils logging (which has been removed).
 
 ## [1.0.0] - 2026-09-09
 ### First Release